+++ example/main.dart
@@ -1,6 +1,6 @@
-import 'package:simple_vcard_parser/simple_vcard_parser.dart';
+//import 'package:simple_vcard_parser/simple_vcard_parser.dart';
 
-//import '../lib/simple_vcard_parser.dart';
+import '../lib/simple_vcard_parser.dart';
 
 String vCardExample40 = '''BEGIN:VCARD
 VERSION:4.0
@@ -13,7 +13,7 @@ TEL;TYPE=work,voice;VALUE=uri:tel:+1-111-555-1212
 TEL;TYPE=home,voice;VALUE=uri:tel:+1-404-555-1212
 ADR;TYPE=WORK;PREF=1;LABEL="100 Waters Edge\nBaytown\, LA 30314\nUnited States of America":;;100 Waters Edge;Baytown;LA;30314;United States of America
 ADR;TYPE=HOME;LABEL="42 Plantation St.\nBaytown\, LA 30314\nUnited States of America":;;42 Plantation St.;Baytown;LA;30314;United States of America
-EMAIL:forrestgump@example.com
+EMAIL;TYPE=INTERNET:forrestgump@example.com
 GENDER:M
 REV:20080424T195243Z
 x-qq:21588891
@@ -25,13 +25,16 @@ void main() {
   print(vc.formattedName); // Forrest Gump
   print(vc.organisation); // Bubba Gump Shrimp Co.
   print(vc.title); //Shrimp Man
-  print(vc.email); //forrestgump@example.com
+  print(vc.typedEmail); // [[forrestgump@example.com, [INTERNET]]]
   print(vc
       .typedTelephone); // [[+1-111-555-1212, [VOICE, WORK]], [+1-404-555-1212, [HOME, VOICE]]]
   print(vc.name); //[Gump, Forrest, , Mr.,]
   print(vc.gender); //M
+  print(vc
+      .typedAddress); // [[[100 Waters Edge, Baytown, LA 30314, United States of America], [WORK]], [[42 Plantation St., Baytown, LA 30314, United States of America], [HOME]]]
+  vc.print_lines(); // Will print all vcard lines without start and end tags
 
-  //getWordOfPrefix() can be used to retrieve values from currently unsupported properties
+  // getWordOfPrefix() can be used to retrieve values from currently unsupported properties
   print(vc.getWordOfPrefix(
       "PHOTO;MEDIATYPE=image/gif:")); //http://www.example.com/dir_photos/my_photo.gif
 }