+++ lib/simple_vcard_parser.dart
@@ -1,20 +1,31 @@
 import 'dart:convert';
 
-class VCardParser {
+class VCard {
   String _vCardString;
   List<String> lines;
   String version;
 
-  VCardParser(vCardString) {
+  VCard(vCardString) {
     this._vCardString = vCardString;
+
     lines = LineSplitter().convert(this._vCardString);
     for (var i = lines.length - 1; i >= 0; i--) {
       if (lines[i].startsWith("BEGIN:VCARD") ||
-          //lines[i].startsWith("END:VCARD") ||
+          lines[i].startsWith("END:VCARD") ||
           lines[i].trim().isEmpty) {
         lines.removeAt(i);
       }
     }
+
+    for (var i = lines.length - 1; i >= 0; i--) {
+      if (!lines[i].startsWith(new RegExp(r'^\S+(:|;)'))) {
+        String tmpLine = lines[i];
+        String prevLine = lines[i - 1];
+        lines[i - 1] = prevLine + ', ' + tmpLine;
+        lines.removeAt(i);
+      }
+    }
+
     version = getWordOfPrefix("VERSION:");
   }
 
@@ -22,52 +33,20 @@ class VCardParser {
     return this._vCardString;
   }
 
-  void print_lines(){
+  void print_lines() {
     String s;
     print('lines #${lines.length}');
-    for(var i = 0; i<lines.length; i++){
+    for (var i = 0; i < lines.length; i++) {
       s = i.toString().padLeft(2, '0');
       print('$s | ${lines[i]}');
     }
   }
 
-  String concatenateLines(int index){
-    /*
-    ADR;GEO="geo:12.3457,78.910";LABEL="Mr. John Q. Public, Esq.\n
-      Mail Drop: TNE QB\n123 Main Street\nAny Town, CA  91921-1234\n
-      U.S.A.":;;123 Main Street;Any Town;CA;91921-1234;U.S.A.
-    ADR;TYPE=HOME:post_office_box;ext_address;address;city;state_or_province;pl
-     z;country
-    ADR;TYPE=HOME:post office box;extended address;address;city;state;postal co
-     de;country
-   */
-    String line='';
-    for(var i = index; i<lines.length; i++){
-      if (index == i){
-        line = line + lines[i];
-      } else {
-        line = line + lines[i].substring(1);
-      }
-
-      if(lines[i+1][0] != ' ') {
-        for(var k = index; k<i+1; k++) {
-          lines.removeAt(index);
-        }
-        return line;
-      }
-    }
-  }
-
   String getWordOfPrefix(String prefix) {
-    //returns a word of a particular prefix from the tokens minus the prefix
+    //returns a word of a particular prefix from the tokens minus the prefix [case insensitive]
     for (var i = 0; i < lines.length; i++) {
-      if (lines[i].toUpperCase().startsWith(prefix)) {
+      if (lines[i].toUpperCase().startsWith(prefix.toUpperCase())) {
         String word = lines[i];
-        if (lines[i+1][0] == ' ') {
-          word = concatenateLines(i);
-        } else {
-          lines.removeAt(i);
-        }
         word = word.substring(prefix.length, word.length);
         return word;
       }
@@ -76,16 +55,16 @@ class VCardParser {
   }
 
   List<String> getWordsOfPrefix(String prefix) {
-    //returns a list of words of a particular prefix from the tokens minus the prefix
+    //returns a list of words of a particular prefix from the tokens minus the prefix [case insensitive]
     List<String> result = List<String>();
 
-    String word;
-    word = getWordOfPrefix(prefix);
-    while(word.isNotEmpty){
-      result.add(word);
-      word = getWordOfPrefix(prefix);
+    for (var i = 0; i < lines.length; i++) {
+      if (lines[i].toUpperCase().startsWith(prefix.toUpperCase())) {
+        String word = lines[i];
+        word = word.substring(prefix.length, word.length);
+        result.add(word);
+      }
     }
-
     return result;
   }
 
@@ -97,7 +76,6 @@ class VCardParser {
     }
   }
 
-
   List<String> get name {
     String _name = getWordOfPrefix("N");
     return _strip(_name).split(';');
@@ -108,12 +86,12 @@ class VCardParser {
     return _strip(_fName);
   }
 
-  String get nickName {
+  String get nickname {
     String _nName = getWordOfPrefix("NICKNAME");
     return _strip(_nName);
   }
 
-  String get birthDay {
+  String get birthday {
     String _bDay = getWordOfPrefix("BDAY");
     return _strip(_bDay);
   }
@@ -146,18 +124,6 @@ class VCardParser {
   String get note {
     String _note = getWordOfPrefix('NOTE');
     return _strip(_note);
-  }
-
-  // GEO:50.858,7.0885          <-- 2.1, 3.0
-  // GEO:geo: 50.858\,7.0885    <-- 4.0
-  String get geo {
-    String _geo;
-    if (version == "2.1" || version == "3.0") {
-      _geo = getWordOfPrefix('GEO');
-    } else if (version == "4.0") {
-      _geo = getWordOfPrefix('GEO:GEO');
-    }
-    return _strip(_geo);
   }
 
   @Deprecated("typedTelephone should be used instead")
@@ -202,7 +168,12 @@ class VCardParser {
         }
       }
 
-      if ( _tel.isNotEmpty) {result.add([_tel, types, ]); }
+      if (_tel.isNotEmpty) {
+        result.add([
+          _tel,
+          types,
+        ]);
+      }
       _tel = '';
       types = [];
     }
@@ -216,50 +187,102 @@ class VCardParser {
     return _strip(_email);
   }
 
-  List<dynamic> get typedEmail =>  typedEmailURL('EMAIL');
-  List<dynamic> get typedURL =>  typedEmailURL('URL');
-  List<dynamic> get typedAdress =>  typedEmailURL('ADR');
+  List<dynamic> get typedEmail => _typedProperty('EMAIL');
+  List<dynamic> get typedURL => _typedProperty('URL');
 
-  List<dynamic> typedEmailURL(String property) {
-    List<String> emailTypes = [
+  List<dynamic> _typedProperty(String property) {
+    // base function for getting typed EMAIL+URL
+    List<String> propertyTypes = [
       'HOME',
+      'INTERNET',
       'WORK',
       'OTHER',
-      'PREF'
     ];
-    List<String> emails;
+    List<String> matches;
     List<String> types = List<String>();
     List<dynamic> result = List<dynamic>();
-    String _email = '';
+    String _res = '';
 
-    emails = getWordsOfPrefix(property);
+    matches = getWordsOfPrefix(property);
 
-    for (String email in emails) {
+    for (String match in matches) {
       try {
-        _email = RegExp(r'(?<=:).+$').firstMatch(email).group(0);
-
+        _res = RegExp(r'(?<=:).+$').firstMatch(match).group(0);
       } catch (e) {
-        _email = '';
+        _res = '';
       }
 
-      for (String type in emailTypes) {
-        if (email.toUpperCase().contains(type)) {
+      for (String type in propertyTypes) {
+        if (match.toUpperCase().contains(type)) {
           types.add(type);
         }
       }
 
-      if ( _email.isNotEmpty) {
-        if(property == 'ADR'){
-          List<String> adress = _email.split(';');
-          result.add([adress, types, ]);
+      if (_res.isNotEmpty) {
+        if (property == 'ADR') {
+          List<String> adress = _res.split(';');
+          result.add([
+            adress,
+            types,
+          ]);
         } else {
-          result.add([_email, types, ]);
+          result.add([
+            _res,
+            types,
+          ]);
         }
       }
-      _email = '';
+      _res = '';
       types = [];
     }
 
     return result;
+  }
+
+  List<dynamic> get typedAddress {
+    List<String> addressTypes = [
+      'HOME',
+      'WORK',
+      'POSTAL',
+      'DOM',
+    ];
+    List<String> addresses;
+    List<String> types = List<String>();
+    List<dynamic> result = List<dynamic>();
+    String _adr = '';
+
+    addresses = getWordsOfPrefix("ADR");
+
+    for (String adr in addresses) {
+      try {
+        if (version == "2.1" || version == "3.0") {
+          _adr = RegExp(r'(?<=(;|:);).+$').firstMatch(adr).group(0);
+        } else if (version == "4.0") {
+          _adr = RegExp(r'(?<=LABEL=").+(?=":;)').firstMatch(adr).group(0);
+        }
+      } catch (e) {
+        _adr = '';
+      }
+
+      if (_adr.startsWith(r';')) {
+        //remove leading semicolon
+        _adr = _adr.substring(1);
+      }
+
+      for (String type in addressTypes) {
+        if (adr.toUpperCase().contains(type)) {
+          types.add(type);
+        }
+      }
+
+      result.add([
+        _adr.split(';'),
+        types
+      ]); //Add splitted adress ( home;street;city -> [home, street, city]) along with its type
+      _adr = '';
+      types = [];
+    }
+
+    return result; // in this format: [[[adr_1_params], [adr_1_types]], [[adr_2_params], [adr_2_types]]]
   }
 }